(db) contact accessing methods for user model
Created the methods 'get_contacts' and 'get_contact_of' for
the 'User' model:

- 'get_contacts' will get all the contacts for the 'User' instance.
- 'get_contact_of' will get all the associations with other 'User'
  instances.

Renamed the foreign key field 'contact' from the 'Contact' model to
'associate', for an attempt to generate better readability. The related
name of 'user' has been renamed to 'contacts' and the related name for
the 'associate' field has been renamed to 'contact_of'.